move error handling into a function that returns unit.
So, uhh...
In order to send diagnostics you need to have returned from `initialize`,
which makes all this stuff go into `initialized`.
initialized though happens returns unit.

so we must handle all the errors with match, there are a few,
and the're all nested.

Then rustfmt fully transforms it into a gorgon, sigh.