be consistent about rebuild dependencies
This fixes issues where if the lexer or parser failed to build,
the parser could parse with the previously built parser tables.

Now we invalidate the parser and lexer tables on lex failure
and parser on parse failure.

Similarly if we edit the lex file, we need to rebuild the parser.
this is not efficient in that we rebuild the lexer whether or not
the lexer changed