incremental report, reparse test input changes
The closures/lifetimes in this commit are a bit of a mess,
and we end up needing to clone a bunch of things, like state tables
etc.

On the bright side, we now report through a channel, rather than collect
results in an array.

and we correctly reparse foo.xyz when foo.xyz changes rather
than *.xyz files, as previous.  when foo.{l,y} changes we do however
reparse all *.xyz files.