add an extension/filetype map users can modify
If your language is 'foobar', but the extension is '.foo'
previously there was no way to stop it from setting the filetype
to 'foo'.

Which could mess with any other things the user has set up
for dealing with 'foobar' filetypes.

Eventually there may be a better way to handle this.